# [GraphBolt] remove buffer in `ItemSampler.__iter__`